implement the comment display
