+++ src/index.js
@@ -1,10 +1,9 @@
 import './styles/style.css';
 import './styles/bootstrap.min.css';
 import './styles/bootstrap-grid.min.css';
-import './bootstrap.bundle.min.js';
-import {
-  renderNavbar, renderShow, addEventToCommentBtn, displayLikes,
-} from './renderDOM.js';
+import renderNavbar from './navbar.js';
+import { addEventToCommentBtn } from './comment';
+import { renderShow, displayLikes } from './renderDOM.js';
 
 document.addEventListener('DOMContentLoaded', async () => {
   await renderNavbar();
